extra functions now in the right places
